Only for linux
Since Windows and Linux handle and take very different shell commands, decided to make a separate script for both.

At some point in the future, I plan on detecting the OS and including both scripts in two if blocks, but for now, I'm working on some more important (imo) features. 

Oh, and also added metadata reading to figure out when a file was modified last and determine its appropriate year folder from that.

It's too difficult to read creation date in Linux, so had to resort to reading the modification date.

The Windows version of this script will be up positively by tomorrow, and that will involve reading creation date rather than modification date.